chore: seperate AsyncStreamProcessor
